checkimages: Add a paramater of maximum number of notifications per user per check
On commons, batch uploads are fairly frequent, we do not wish to flood
a user's email inboxes just because they forgot to add a few characters
to the file description pages. By adding this limit, the severeness of
email flooding can be greatly reduced, and a few talk page notifications
is usually enough to let batch uploaders know about their issue.

Bug: T157728
Change-Id: I9e25b35d42af3c73c581c6d05a6a4a8dc8ef62c6