Revert "Set minimum requirement of websocket-client"
Pywikibot does not depend on a particular version of websocket-client;
in fact it does not even interact directly with its API.  Stating a
requirement on a particular version because some bug was fixed there
has the disadvantage that it disallows previous versions to be used
where the bug did not occur (< 0.27.0), and it also prevents
backporting fixes (0.27.1 or 0.27.0~local_patch).

This change reverts commit b96e7f57905751c08a0b5c5197d47dff3bcfe182.

Change-Id: Ie2ba76cf0280dc095539c6b2648adc5e67085782