Fix logging in python3 when deprecated_args decorator is used
Remove check on record path, which was failing in python 3.

Remove special treatment between python 2 and 3, leaving the default
behaviour for each version.

Logging out put is slightly different in python 3.6 with respect to the
others, as record.pathname is warnings.py.

Leave indication of original caller module:linenno in message.

Bug: T159077
Change-Id: I0c13551406e709d0082874d383962fd514dcae99