[Bug] Inject correct dependencies to SparqlQuery
I057335a broke most (third-party) usages of WikidataSPARQLPageGenerator
because only providing the endpoint to the SparqlQuery interface now
throws an error.

The user can now just provide the repository if it's enough,
otherwise they need to provide the mandatory dependencies as well. The
decision about whether it's enough is now made inside SparqlQuery.

Change-Id: I2c2c383b39b3684d8aaed0375ab4253168c135e3