tools_tests.py: Take care of windows line endings
Depending on git configuration, line-ending of checked-out files
may be "\r\n" instead of "\n" for windows users. This causes the
hash values of files to change and tests will fail. Add hash
values for windows line-endings.

Replace b'\r\n' with b'\n' in the content of uncompressed files.

Change-Id: I5babeb7b3ea5d1800449e37a8a2b2c358eec663a