Add support for tabular-data Wikibase data type
Introduce the WbTabularData class to provide handling of tabular-data
values and do some basic validation on indata.

There are significant overlaps with WbGeoShapes as they only differ
in which repository they (can) use and the required ending of the
page name (.tab rather than .map). As such they should probably be
rebuilt on a shared base at a later stage.

Bug: T163981
Change-Id: I3f15d9732f7bbb552b1d000530d295fccca2fd5f