Clean up issue with _WbDataPage
The first merged patch (https://gerrit.wikimedia.org/r/351714)
was broken and a quick fix (https://gerrit.wikimedia.org/r/355812)
was introduced to stop the tests from crashing.

This patch cleans up the issue in the original patch.

I removed the None default for site since the site value should
always be passed to this function and since the function has not
been in production long enough to warrent allowing it to not have
the site passed simply for backwards compatibility.

Bug: T166362
Change-Id: If2790daeb2c11ff8622f95a8c4cec3a9700d667c