[IMPR] Make socket_timeout recalculation reusable
On older installations of requests it's not possible to use two timeout values
but newer versions may have two: one for the connection one for read timeout.

To ensure the right value depending on the requests version there was a fix
inside http.py; this should be moved to the upper library place to be
reusable by other library scripts like rcstreams or eventstreams.

config2.py:
- Since requests > 2.4.1 is required with T134647 for pwb now
  the default socket_timeout may be a tuple again. Use new values for it.
- fix the socket_timeout to a single value dependig on
  requests version but use the higher of the two values.
  Remove warning because labs is still using older requests version 2.2.1.

- http.py:
  remove the socket_timeout fixer which was transfered to config2.py.

Bug: T166539
Change-Id: I598ac9fa2fa4f237b68d44bd8cf3c587d5768bd1