Allow pywikibot to authenticate against a private wiki.
Currently, pywikibot cannot authenticate if MediaWiki is configured
to require logins in order to read. This patch takes an approach
suggested by user @zhuyifei1999 in ticket T153903. It tells
pywikibot not to require the parameter "meta=userinfo" if
'tokens' in present in parameter 'meta' and 'login' is present in
parameter 'type'.

Bug: T153903
Change-Id: I55768b735dbddcf35dbb35b94e9ba493621ff51b