Add ability for Pywikibot to thank Flow posts
This change adds a thank method to Post, which will allow bots to
thank the creators of Flow posts. A property was added to Post
which returns a User object corresponding to the creator of that
post. Two tests, adapted from the revision thanking tests, have
been added and pass locally.

Bug: T135411
Change-Id: Ie629aaa46aa007a39e6d5dce6400504bf73e5980