Handle non-items in ItemClaimFilterPageGenerator
Now if there are non-items in the generator, we try
to find the item first and then filter on it. However,
we have to abandon preloading since we can no longer
assert that we are working on a list of items.

I have also made the type check weaker, so that other
types of entities are supported as well.

Bug: T137842
Bug: T160396
Bug: T170991
Change-Id: I06851f18d85228d5db77cdec5ed1e07fcda32f7a