[test] Replace expected_failure_if with skipIf decorator
unittest's expected failure decorators are handy when a bug was
identified in the code but wasn't fixed yet in the current release.
The different behavior of py2 and py3 related to string elements
which aren't string constants cannot be solved because it is an
upstream problem and part of the release difference issue.

Therefore we shouldn't use that expected failure reminder but just
skip the test for py3.

Bug: T178528
Change-Id: I40fce5d7a4605fcc1b2cdd7b811c31eef53d4a2c