[compat] do not actively support compat module
core branch is now 11 years old and decommission of compat was two years
ago. The compat module was created in 2012 and older catlib and userlib
where merged into that folder in 2014.

The framework was never using this compatibility module and bot operators
get a deprecation warning for their private scripts when the module is
used. The scripts should be updated accordingly.

- remove their references from autogenerated docs to not actively support it
- update compat2core.py maintenance script

detached from I32cd9ca2d1e

Bug: T183085
Change-Id: I9d61e1eb0d26e570406b14ee28b0cd86a2a49871