[compat] Remove compat module
core branch is now 12 years old and decommission of compat was closed
four years ago. The compat module was created in 2012 and older
catlib and userlib where merged into that folder in 2014.

The framework is not using this compatibility modul anymore and bot
operators got a future warning for their private scripts when the
module is used.

The README-conversion.txt as well as the maintenance/compat2core.py
script was updated years ago not to use this framework part anymore.
There are core framework equivalent parts for userlib, catlib and the
query module and the behaviour of the old parts was never tested.

Anyway the compat implementation of User, Categrory and query is
very different from the core implementation and it is time to take
the next step.

- drop compat library module
- remove references from docs
- remove compat folder from pwb sys path
- update docs

Bug: T183085
Change-Id: I32cd9ca2d1ef32a25f741fe46160317cb649c4ef