[IMPR] Introduce setup and teardown methods
setup and teardown methods can be used to initalize and cleanup the bot.
setup is called in when the bot.run starts working whereas teardown is called
from exit method. This is much more clearer than doing this inside a derived
run method.

setup allows some inital work like reading from live wiki or io stuff
which might be unwanted in the initializer of the instance. teardown allows
some cleanups like saving content to a file when the bots stops running and
there is no need to derive the exit method.

Change-Id: Id18281feb67be8a39f79e409a752f9502e736850