CategorizedPageGenerator: Use startprefix parameter of category.articles
The previous method of passing `start` to `startsort` was wrong. It does not
work as expected because `startsort` needs to be a binary string.[1]

For older versions of MediaWiki implement a workaround by iterating over
all category.articles. This is inefficient, but no better way found.

Remove `kwargs['sortby'] = 'sortkey'`, 'sortkey' is API's default.

[1]: https://www.mediawiki.org/wiki/API:Categorymembers

Bug: T74101
Bug: T143120
Change-Id: I7eece66e258659f84479fbdd1094a8df9ee69479