page.Category.articles: Fix the equivalant of startFrom in core
The way `startFrom` works in compat is equal to `startprefix` in core.[1]
`startsort`, which is deprecated in MW 1.24, needs to be a binary string,[2]
otherwise the user will get the wrong results because of the wrong start.

This means that compat's `startFrom` has no equivalant in core if the
MediaWiki version is less than 1.18.

[1]:
https://phabricator.wikimedia.org/diffusion/PWBO/browse/master/catlib.py;dfd2f011fa9614e2131743e761f1dbafe7f2f1a0$327
[2]: https://www.mediawiki.org/wiki/API:Categorymembers

Change-Id: I10620e82d2d0792579889199ffa5e94a0cce0d07