Allow archivebot to archive even restricted pages
The archive templates were recently added to templates which restricts
page to be edited, but it made archivebot not archive at all.

This patch excludes archive templates from restrictions and introduces a
method to filter out archive templates if archivebot is there.

TODO: This patch could be possibly expanded to some form of force
categories like:

    page.save(force='always') x 'archive' x None

Change-Id: Iab249fd83f4f071b4694eba8d0b7d977ed22c3f8