Sphinx autodoc workaround for Family classproperty-s
Sphinx uses getattr() on the class in order to get the attribute.
This works for normal instance properties, since it gets the property
instance itself, but for classproperties, it gets whatever the
classproperty code returns. This not only loses all documentation
the classproperty code have, and can potentially cause side effects.

Other methods to workaround the issue are described in:
https://stackoverflow.com/q/22357961. This includes subclassing the
return value (unknown consequences to user), metaclass property
(not extensible), proxy object (error-prone and ugly),and indirect
get method (breaks backwards-compatibility entirely). This workaround
makes Sphinx uses a custom getattr we defined in docs/conf.py for
getting attributes of Family subclass, where it iterates through the
MRO looking for a definition. If it's defined as a classproperty,
it returns the classproperty instance instead; else it falls back
to sphinx's safe_getattr().

Other changes in this patch include:
* make all families defined in family.py abstract and
  un-instantiatable.
* make classproperty fetch its docstring from its wrapped function
* make docs/conf.py have its `from script.` import after insert to
  sys.path so that it is easier to run sphinx locally.
* make SubdomainFamily.codes no longer treat empty languages_by_size
  as valid, because empty languages_by_size is always defined via
  MRO and `hasattr(cls, 'languages_by_size')` is never False. This
  also prevents SubdomainFamily.codes being accessed directly.

Change-Id: I790580b8d5d9e6349a26ab52c8859a7148441bff