shell.py: call `sys.__interactivehook__()` if available
This pactch makes shell.py acts more similar to the built-in
interactive shell, where Python native code calls the hook. The
hook may be user-defined, or be the default implementation
defined by the site module of CPython 3.4+ Standard Library. The
site module is automatically imported if -S is not specified to
the Python interpreter.
The default implementation of the hook does all the magic of
loading readline, saving history, and tab-completion.

Unless the user defined their own sys.__interactivehook__, which
may include other features and/or exclude readline, and which we
honor in this patch, by default, this patch only works in CPython
3.4 or newer, after the fix of Python issue #5845 [1], and only on
Python installs where GNU readline is compiled into the CPython
interpreter binary. Other Python implementations, such as
IronPython, may also include their implementations of the hook,
and we try our best to simulate their built-in shell, but it is
not our main concern.
* In Windows, Python uses the ReadConsoleW() Windows API, which
  has some cmd-like history functionality built-in. It is able to
  use the history of a previous invokation of Python as long as
  the terminal window stays open.
* In older versions of Python, sys.__interactivehook__ is not
  defined by site.
* In newer versions of Python, but without readline compiled,
  interactivehook will attempt to import readline but silently fail.
In these cases, you get the plain old interactive prompt, as if
this patch never existed.
In any case, this patch shall attempt to make the behavior of
shell.py closer to what would be provided if one invokes the
built-in Python interactive shell directly.

[1] https://bugs.python.org/issue5845

Bug: T196168
Change-Id: I410c9717370cff70fa441f06500625956e8596fc