Fix handling of API continuation in PropertyGenerator
test_preload_langlinks_count was marked as allowed_failure. It relies
on site.preloadpages, which relies on PropertyGenerator, which had
a bug in it.

The issue was that the default implementation of QueryGenerator
returns incomplete results: Sometimes the API splits long pageprops
into several API requests. But QueryGenerator returns the result as
soon it gets it, without waiting for it to complete in the next
request. In other words, PropertyGenerator used to yield more than
expected, but each yield had less than expected fields. Therefore
site.preloadpages returned more pages than expected, but the
preloaded data on each page could be incomplete. There are many other
function that rely on PropertyGenerator and might be effected
by the bug, but I have not performed any test to confirm that.

I don't know if other generators also have a similar bug or not. This
patch only tries to fix PropertyGenerator. If such issues are found
in other generators as well, then it might make sense to move some of
the newly introduced PropertyGenerator methods into QueryGenerator.

api.py:
  - QueryGenerator:
    Do a refactoring to move a small part of `__iter__` into a
    separate function (_extract_results). We can override and extend
    this function later in the PropertyGenerator subclass.
  - PropertyGenerator:
    Override `__iter__` and `_extract_results` in a way to prevent
    yielding premature results. A result is considered fully fetched
    if it is not found in the next request results.

site_tests.py:
  - Remove the allowed_failure decorator and reenable the test.
  - Silence the output.

Bug: T196876
Change-Id: I4ed74ab3af6d320242beaef1a55f20dc489fe29b