family: read `__init__` from `__dict__` and deprecate
deprecated wrapper expect a pure function, but in Python 2.7,
reading the function as a class attribute returns a method bound to
the class (unbound method), which cannot have arbitrary attributes
assigned, and the deprecated wrapper errors out. Reading from
__dict__ should return the actual function regardless of the Python
version used. The existence of the __init__ key is verified prior
the the execution.

Bug: T197529
Change-Id: I2077c6b0fa95c3dc2da0f8c1c5e99d12372ebcb9