Disable the need_extension('WikibaseClient') decorator on unconnected_pages
Pywikibot caches the result of `CachedRequest`s and it seems that AppVeyor
is retaining that cache between project builds. This could also be true for
bot users.

This patch temporarily disables the need_extension decorator so that the
check is not performed. It should be OK to reenable it after 30 days when
the cache expires.

Bug: T198454
Change-Id: If64d7e48c71a79a3ff2ca63c71e1ccd9446cc96b