[cleanup] Remove unsupported removeImage and placeImage Page methods
removeImage and placeImage Page methods where never part of the
rewrite/core release. A deprecation warning is shown if the one
tries to use it but the method does just nothing.

Remove this part of code after this long time (since 27.02.08),
because it seems better to raise an attribute error instead of
having a wrong behaviour for old scripts.

Bug: T106121
Change-Id: Ieb1d6a0cad06732699fff0aac01e91c358538540