transliteration.transliterator.__init__: Use a literal dictionary
According to codeclimate transliterator.__init__ has a very high
cognitive complexity due to its long length.[1]

This patch fixes that by using a dict literal defined as a global
instead of gradually building the transliteration dict on instantiation.

A few duplicate keys were removed in a way that the result has not changed
at all.

[1]: https://codeclimate.com/github/wikimedia/pywikibot/issues

Change-Id: Iecc0607c9416f7f345877fb36a2be9ee30d4412c