Use a key for filter_unique where appropriate
filter_unique can potentially consume a lot of memory since it stores
returned objects in a set if no `key` is provided. (T199615)

This patch adds an appropriate `key` to filter_unique calls wherever
the aforementioned issue might be a problem.

filter_unique has several usages:
  - tools_tests.py:
    We don't have any memory issues here.
  - site.py:
    No change is needed. The only usage of filter_unique is to filter
    unique pageids which do not consume much memory.
  - pagegenerators.py, scripts/template.py, scripts/nowcommons.py:
    filter_unique is applied to page generators. Use
    '{}:{}:{}'.format(*page._cmpkey()) as the key.

This patch-set is being proposed as an alternative to [1] that avoids
making `hash` the default key of `filter_unique`.

[1]: https://gerrit.wikimedia.org/r/c/pywikibot/core/+/445854/

Bug: T199615
Change-Id: Ia701e9f8bece713174ea4bccd104c0ae25ff5c1d