Introduce a new APISite property: mw_version
APISite.version() is a widely used method. In most of its usages the
version string is converted to a MediaWikiVersion object.

To improve readability and ease of use, introduce a new property,
mw_version, which returns a MediaWikiVersion object and caches
the result.

Use the new property throughout the codebase.

Deprecate pywikibot.MediaWikiVersion in favour of
pywikibot.tools.MediaWikiVersion.

Change-Id: I75de443b24d2eb478724c416db8f2d734b6bb806