Fix type hints for *args and **kwargs
First of all, *args is not a list, it's a tuple. In addition to that,
and according to PEP 484[1], the type hint for *args should be the type
of its elements, otherwise it's obvious that the type of *args is
a tuple. Similarly, for **kwargs the type hint should mention the
dict's value types.

Use `typing.Sequence[T]` or `typing.Iterable[T]` for args (without *).

[1]:
https://www.python.org/dev/peps/pep-0484/#arbitrary-argument-lists-and-default-argument-values

Change-Id: I7f0db256b52714c5e46db0a9f609831bee74ac53