Removed unused vars/parameters from daemonize.py and data/api.py
According to vulture the code was not being used. I removed it and ran tests. As far as I can see
I did not break anything. But it is my first commit so it might be good if someone else ran tests
on it.
I also removed a superfluous error from data/api.py and removed the per-file-ignores that ignored
the error.

Bug: T203395
Change-Id: Id75da257c2f3e3be43b98e749f027258a1a576e1