atexit: Stop polluting stdout with sys.last_type
Some people (eg. me) use more of pywikibot as a library, and in some
uses, like text stream processors, stdio pollution is simply not
tolerable.

Besides, sys.last_type should already be given, at exit time, by
sys.excepthook, to stderr. There is no reason whatsoever the cause
of exit should be duplicated to stdout when it's already in stderr.
In the extreme case that a user actually rely on this bug feature
(I can't imagine why but it's possible, given this code has existed
for so long that I finally hit some use case that can't tolerate
this anymore), it's very easy for them to register their own atexit
handler; on the other hand, it's very hard for me to silent this
atexit handler without directly removing the line.

This is so bad that we even had to create an exception for T001 in
tox.ini without any explanation why this print must exist in the
first place.

Oh and in case anyone wonders, this print was added almost 5 years
ago in 4b09153, and in that diff it looks more like for debugging
or testing of that patch rather than something critical.

Considering that the logged messages at exit may be the same (except
the log level of 'Closing network session.'), the exception type is
now logged in a more meaningful message in critical log level and
the original 'Closing network session.' is now always info level,
since it is the exception, not the network session being closed,
that is of interest to the user.

Change-Id: I83ce680f7e4eeb2e18fcfc60db70e88a8c668538