tox: stop setting a default command
The default testenv defines a default command which might have
unattended effect.  A current example is the envlist defining
'doctest-py27' and 'doctest-py34', they are not matched by the section
[testenv:doctest] and end runnining the default command.

Prefix py27 and py34 with 'test-', tox will thus generate a factor named
'test' which we can use to define a command for.

Change-Id: Ic907ea89d8b6d96efe1bf3149cc637c9d33cc627