mysql: drop support for MySQLdb module
The PyMySql module has been the default since 72f072f0c and is the
prefered python module.  Its advantage over MySQLdb is the module is
pure python and does not depend on having the mysql client to be
installed on the host. It is thus easier to install for end users.

Remove left over legacy support of MySQLdb.
Add a note in HISTORY.rst as a hint.

Note that the code is only used by pagegenerators.py

Bug: T89976
Bug: T243154
Change-Id: Iaa295b27269d40ae6f157635044493566ce39b18