[IMPR] Start the script if its name is similar to the given script name.
With  I0fb163eef misspelled scripts are
offered to the operator to choose the right one.

If only one script is found, start this without asking the bot owner
but give him a notice and wait a bit to let it be cancelled.

Note: I0fb163eef must be merged first

Bug: T217195
Change-Id: I4b4873e7172d98f1c7d91ef76eea8d6f775809ee