[tests] Give a better failure message for paraminfo_tests
Currently we get an assertion error listing two sets when the first set
is not greater or equal than the second.

Now calculate the difference which values are in the second and not
in the first set as expected.

This is an improvement to detect the missing parameter in the first.

Bug: T218506
Change-Id: Ibcdffd5050c136805d48bf5dddd67011d81d84d7