Improve title patterns of WikibasePage extensions
Having ^ at the beginning of a pattern a text is
re.match'd against is pointless. Without ^$,
users can use the patterns for arbitrary string
matching (though we need to refactor the -1 title
handling first). Python 3.4 has a dedicated function
for matching without ^ and $ (re.fullmatch) but
we are still Python 2.7 compatible.

Change-Id: Id0a9b2fc74bd27873f100294c20d1aa6b587fe20