[bugfix] Enable all families are tested in TestFamilyUrlRegex.test_each_family
With skipTest the whole test is skipped if the condition was reached.
In this case all tests after 'lyricwiki' was found in config.family_files
where droppend and btw the order of the config.family_files.keys seems
not deterministic.

With this patch a subTest is used. subTest was introduced with Python 3.4.
After we annouced to drop Python 2 it is not necessary to backport this
context manager or find another solution. Just skip this test for older
Pythons.

Bug: T223367
Change-Id: Ief76609cfbdc976a34f5161373015ab25343860f