[IMPR] Provide FutureWarning with deprecating methods
FutureWarning is a warnings about deprecated features when it is
intended to be seen by end users of the applications.
(See also PEP 565)

On pywikibot each FutureWarning is only shown once when running a script.

Enable deprecating methods and decorators to show FutureWarnings
instead of hidden DeprecationWarnings. This can be used if old
deprecated code will be upcoming removed.

Change-Id: I9a79c7458b2a8f1249088c160f983e428835840f