[IMPR] add a new ConfigParserBot class
ConfigParserBot enables to set options from the scripts.ini setting file.
The setting section for the script is identified by the module name.
All settings in the settings file must be listed in the availabelOptions
dictionary of the bot. The type of the value is identified by the
availabelOptions default.

The script's settings is like this:

[add_text]
  summary = Bot: Aggiungo template Categorizzare

[shell]
  always=true

Bug: T223778
Change-Id: I601a5bc055239a7848c897b4651ed22f80af2ff0