refactor(config2): remove the unimplemented proxy variable
Users can use environment variables to setup a proxy as mentioned in [1].
Users who already have this unused variable in their user-config.py file
will now recieve the following warning:

"""
UserWarning: Configuration variable "proxy" is defined in your \
user-config.py but unknown. It can be a misspelled one or a variable
that is no longer supported.
"""

[1]:
https://www.mediawiki.org/w/index.php?title=Manual:Pywikibot/Installation&diff=3318238&oldid=3315748

Bug: T111752
Change-Id: Ife62e94d13fc6c33e8a19a49bf855f26cbb49a65