test(test_family_load_valid): do not care if languages_by_size is empty
I've created a few custom family files, e.g. for wpbeta, but they fail
this test because the autogenerated family files does not have a
languages_by_size attribute.

We have fixed and changed this part of test several times during these
years. I think it is not worth maintaning and is safe to remove.

Change-Id: Icffca034d26a82d9114da6c0668ac32bc07c060e