test(utils.py): redesign the timeout workflow
The implentation of timeout is complex and confusingly just returns
empty strings if a timeout occurs, causing T224364.
Use a new implementation for Python 2 which raises TimeoutError and
use the default one in Python 3.

In this design, infinite timeouts are indicated by None, not 0.
Change the script_tests accordingly.

Increase the timeout of test_one_similar_script to 10 seconds.
It occasionally failed for me with the previous 6s timeout.

Bug: T224364
Change-Id: Icc69d3df8ca41b86061f40fcaef0da7fa46d494e