Deprecate the ability to login with a secondary sysop account
Once this patch is merged, any part of the code that used to rely on
a secondary sysop account (through the use of sysopnames) should stop
working. I am intentioally postponing the removal of sysopnames from
other parts of code (specifically, pywikibot/__init__.py) to a later
time because this "agile" patch will give us the opportunity of
receiving feedback from users without waiting for a much larger change
to be reviewed.

Bug: T71283
Change-Id: I577ed1b2128856920a8e2a531a53df9fa0e8c5c5