Replace "-" with " " when comparing titles
`BaseSite.sametitle` has been giving me trouble as I've been working with (wikiHow)[www.wikihow.com]. Working with the article https://www.wikihow.com/Be-Cyber-Goth, for example, the titles `"Be-Cyber-Goth"` and `"Be Cyber Goth"` are compared and are judged not to be equal. This PR replaces `"-"` with `" "` (which is also how `"_"` is handled), which I think is the correct behavior. Please let me know if you'd like more information.