[bugfix] Fix TestDeletionBot missing user
After replacement of "test" family with "wikipedia" family in tests,
this error is thrown. Load user for the test to make it pass.

It seems this is caused by the fact `delete` now checks for user rights,
which needs user to be provided. Weird enough test family worked well,
there was some issue with test family user in tests as well.

Bug: T244027
Change-Id: Ic084b025ac2ec8d0457ed4fb0d942cc17ca16024