[IMPR] Use UserDict for LogEntry
A UserDict holds its data in a data attribute which is a dict.
The dict can be accessed directly or via data attribute. This
enables to access all LogEntry entries directly with its key
even there is no corresponding method to retrieve the data.
The access via the data attribute still works.

- Remove the LogDict helper dict which is no longer used.
- Copy the one and only __missing__ method from LogDict
  to LogEntry class which works like a dict now.
- Remove data._type attribute which was used by LogDict
- Shortcut each getting data; the data attribute is no
  longer necessary.
- Remove __ne__ method which just returns the opposite
  of __eq__; this is done by default.
- Remove logid, pageid, ns, type, action, user, comment
  methods which just returns the corresponding data and
  replace them by __getattr__.
- Some tests added to verify that LogEntry.data[item] is
  equal to LogEntry[item] managed by UserDict.

Change-Id: Ic04f63150fd29123b2459dbf291d79e71fb95bf8