[IMPR] move page.py to its own folder
page.py is one of the biggest framework files with 253 KB disk space
and 6800 lines of code. It contains 17 classes and 5 functions.
This should be splitted into smaller parts for readability and
maintainability. Moving page.py to page/__init__.py is the first step.
See I8495a6ae7 for a splitted example.

Change-Id: I01a2c5701d6e094efc20dc86706e34c1ef7cbaf2