[cleanup] Remove old compat line_separator config variable
line_separator was introduced due to different behaviour when page
content was retrieved by api and screen scraping. Therefore the
LS at core is '\n' whereas at compat it was '\r\n'. That config
variable was mainly for having scripts in sync for the both
framework variants compat and core; the scripts where mainly the
same but the LS was different in config files.

Replace all occurences with '\n'

Change-Id: Ia7fe8d56c52a8430385dc4dc9014b5fad758c27e