[cleanup] instantiate a CosmeticChangesToolkit by passing a page
The CosmeticChangesToolkit interface was expanded step by step
but the parameters were kept due to historical reason even though
most of them can be derived from page object directly.

- instantiate the CosmeticChangesToolkit from a page object
- only allow keyword arguments except for page parameter
- update doc of initializer
- show a FutureWarning if a CosmeticChangesToolkit is still created
  by site, namespace and pageTitle
- raise a TypeError if a CosmeticChangesToolkit is created by a page
  object and namespace or pageTitle are still given
- show FutureWarning to deprecate from_page method
- update CosmeticChangesToolkit calls in page library
  and cosmetic_changes.py

A FutureWarning will be shown in favour of DeprecationWarning to
switch the code by passing a page very shortly. This gives us the
ability to get and set the text from and to the page object
directly if applicable.

Change-Id: Ic98fe7a01a221972c6474871679769a98240d99e