[IMPR] Unify and extend formats for setting sitelinks
Both list and dict of sitelinks will be accepted. A sitelink
can be represented with JSON as a dict, Page object, Baselink
object or a string (if the input is a dictionary).

Although SiteLink objects inherit from Baselink, they get
specific handling, so that there is convenient way to set
a sitelink with badges.

Bug: T225863
Bug: T251512
Change-Id: Ie560b8dd72201f4e71d524af326dac8067774ad2