[cleanup] Always use mwparserfromhell if present
use_mwparserfromhell is True by default and mwparserfromhell will
be used in that case if the package is installed.

The extract_templates_and_params_regex function is also available
if mwparserfromhell should not be used in rare cases. Note that
this fallback does not work for deeply nested templates (> 3).

Update tests

Change-Id: Id055618106eec13957c0038210cbabf19771ad5b