[FEAT] Users might want to know only the count of their watchlist
In a case where we have a large list of watchlist items and the terminal
can't scroll to the top due to buffer overflow, users might want to know
just the number of items/pages in their watchlist.

This feature introduces -count and -count:all options to return total number
of pages in the watchlist the bot has access to (for that wiki or all wikis
respectively).

In addition, to avoid a run-time error I encountered that says:
“RuntimeError: dictionary changed size during iteration", I had to make
a copy of the keys of the config usernames via `list(config.usernames)`.

Change-Id: Ie535f12a51f9ebf3b1b2dd11ef4d3bf544c42929