Do not exclude userscripts directory from pycodestyles
This directory is empty by default. Whether the user choses to put
codes in it that are PEP8 compliant or not is up to the user. So,
the choice of whether to exclude this directory from pycodestyles
should also be up to the user.

This reverts Ic97089c806a5eacc332f30750431adf67be3cb56

Additionally, do not exclude user-config.py from the flake8 checks
for the same reason.

Bug: T253873
Change-Id: Ied5f8d5eca71d510269336757bab9f6f7e3dd081