[IMPR] Wait for _putthread is done in BaseBot.exit()
Re-implement waiting for for _putthread is done in BaseBot.exit()
but use stopme() function instead of _flush.

The reason for the revert of 87cfda61cd98 was that async put was blocked:
With _flush a tuple containing None was added to the page_put_queue and
the async_manager function leaved the loop and no longer processed any
queue items which were put later by async_request. This means all
following wheren't pocessed anymore i.e. no page was written to site.
in result the page_put_queue increases up to its max_queue_size and blocks
after that including KeyboardInterrupt wasn't funtional.

This problem only occures if the bot itself runs in a loop.

This partly reverts commit 4291f36321606 but re-implements most of the
87cfda61cd98 patch except that _flush was replaced by stopme,

Change-Id: Ic26928eafdf19844c950f40993073a4e6b9dd687