page.botMayEdit: Use site.username() instead of site.user()
site.user() only returns a username when the bot is logged in, according
to site.logged_in(). However, bots using OAUTH are not considered to be
logged in, so site.user() returns None. This causes OAUTH bots to ignore
the username lists in {{bots}}, which is highly undesireble.

site.username() will return the username that the bot would use if it
were currently logged-in, regardless of if it is or not. If no username
is specified at all, it returns an empty string instead of None. This
will still result in botMayEdit ignoring the username list when no
username is found.

Bug: T153541
Change-Id: Icf9e6f2c7181ddcf908475b075a5eff32a8abf78