[4.0] Remove Python 2 dependencies
With this release Python 2 and Python 3.4 is dropped. This patch only
disables Python 2 tests, requirements are Python 3.5+ only and setup.py
has been changed that Python 3.5+ is needed for site package
installation. The pwb.py wrapper and the framework still works with
Python 2 as long as the installed packages are compatible.

- update doc in docs/requirements-py3.txt
- disable Python 2 tests for Appveyor
- disable Python 2 tests for Travis
- disable Python 2 fasttests and update tox.ini
- use Python 3.7 for py27 until integration-config-tox-docker was
  updated
- remove outdated requirements in requirements.txt
- remove outdated requirements in dev-requirements.txt
- delete requests-requirements.txt which is no longer needed

setup.py:
- Python 3.5+ is needed to run setup.py
- remove outdated dependencies
- remove Python 2 code parts
- Update python_requires and classifiers

Bug: T239542
Bug: T213287
Change-Id: Ia4640ce891d6e023da2cd722e175cbd737dee0bb